backend/pyproject.toml
[project]
name = "fantasy-manager-backend"
version = "0.1.0"
description = "Fantasy Manager API backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=4.0,<4.1",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "email-validator>=2.0",
    "apscheduler>=3.10,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "factory-boy>=3.3",
    "ruff>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]

[tool.ruff.lint.per-file-ignores]
"app/api/**" = ["B008"]  # Allow Depends() in function defaults (FastAPI pattern)
"app/auth/**" = ["B008"]  # Allow Depends() in function defaults (FastAPI pattern)

[tool.ruff.format]
quote-style = "double"

[dependency-groups]
dev = [
    "respx>=0.22.0",
]
